feat: 추상클래스, Interface 과제 2
