+++ src/main/java/com/survivalcoding/asset/Asset.java
@@ -1,0 +1,27 @@
+package com.survivalcoding.asset;
+
+public abstract class Asset {
+    private String name;
+    private int price;
+
+    public Asset(String name, int price) {
+        this.name = name;
+        this.price = price;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getPrice() {
+        return price;
+    }
+
+    public void setPrice(int price) {
+        this.price = price;
+    }
+}
