+++ src/main/java/com/survivalcoding/asset/TangibleAsset.java
@@ -1,30 +1,11 @@
 package com.survivalcoding.asset;
 
-public abstract class TangibleAsset {
-    private String name;
-    private int price;
+public abstract class TangibleAsset extends Asset {
     private String color;
 
     public TangibleAsset(String name, int price, String color) {
-        this.name = name;
-        this.price = price;
+        super(name, price);
         this.color = color;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public void setPrice(int price) {
-        this.price = price;
     }
 
     public String getColor() {
